requirements.txt
pytest==7.4.3
requests==2.31.0
faker==20.1.0
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.5.0